Restructure code: move managers and controllers into their own namespace
